+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - イベント作成処理を最適化
 - イベント検索条件拡張
+- イベントのoffers.itemOffered.serviceTypeを最適化
 
 ### Deprecated